requirements.txt
numpy
pandas
scikit-learn
streamlit
Pillow
yfinance
matplotlib
scipy
keras